+++ cross/transformations/categorical_features/categorical_enconding.py
@@ -1,14 +1,6 @@
+import category_encoders as ce
 import pandas as pd
-from category_encoders import BinaryEncoder
-from category_encoders.leave_one_out import LeaveOneOutEncoder
-from category_encoders.rankhot import RankHotEncoder
 from sklearn.base import BaseEstimator, TransformerMixin
-from sklearn.preprocessing import (
-    LabelEncoder,
-    OneHotEncoder,
-    OrdinalEncoder,
-    TargetEncoder,
-)
 
 
 class CategoricalEncoding(BaseEstimator, TransformerMixin):
@@ -26,12 +18,10 @@ class CategoricalEncoding(BaseEstimator, TransformerMixin):
     def set_params(self, **params):
         for key, value in params.items():
             setattr(self, key, value)
-
         return self
 
     def fit(self, X, y=None):
         self._encoders = {}
-
         X = X.copy()
 
         for column, transformation in self.encodings_options.items():
@@ -40,81 +30,94 @@ class CategoricalEncoding(BaseEstimator, TransformerMixin):
         return self
 
     def _fit_encoder(self, X, y, column, transformation):
-        if transformation == "label":
-            self._encoders[column] = LabelEncoder().fit(X[column])
+        encoder_classes = {
+            "backward_diff": ce.BackwardDifferenceEncoder,
+            "basen": ce.BaseNEncoder,
+            "binary": ce.BinaryEncoder,
+            "catboost": ce.CatBoostEncoder,
+            "count": lambda: X[column].value_counts().to_dict(),
+            "dummy": lambda: ce.OneHotEncoder(drop_invariant=True),
+            "glmm": ce.GLMMEncoder,
+            "gray": ce.GrayEncoder,
+            "hashing": ce.HashingEncoder,
+            "helmert": ce.HelmertEncoder,
+            "james_stein": ce.JamesSteinEncoder,
+            "label": ce.OrdinalEncoder,
+            "loo": lambda: ce.LeaveOneOutEncoder() if y is not None else None,
+            "m_estimate": ce.MEstimateEncoder,
+            "onehot": ce.OneHotEncoder,
+            "ordinal": lambda: ce.OrdinalEncoder(
+                mapping=[
+                    {
+                        "col": column,
+                        "mapping": {
+                            k: v for v, k in enumerate(self.ordinal_orders[column])
+                        },
+                    }
+                ]
+            ),
+            "polynomial": ce.PolynomialEncoder,
+            "quantile": ce.QuantileEncoder,
+            "rankhot": ce.RankHotEncoder,
+            "sum": ce.SumEncoder,
+            "target": lambda: ce.TargetEncoder() if y is not None else None,
+            "woe": ce.WOEEncoder,
+        }
 
-        elif transformation == "ordinal":
-            self._encoders[column] = OrdinalEncoder(
-                categories=[self.ordinal_orders[column]]
-            ).fit(X[[column]])
-
-        elif transformation in ["onehot", "dummy"]:
-            drop = "first" if transformation == "dummy" else None
-            self._encoders[column] = OneHotEncoder(
-                sparse_output=False, handle_unknown="ignore", drop=drop
-            ).fit(X[[column]])
-
-        elif transformation == "binary":
-            self._encoders[column] = BinaryEncoder(return_df=False).fit(X[[column]])
-
-        elif transformation == "target" and y is not None:
-            self._encoders[column] = TargetEncoder(smooth=0).fit(X[[column]], y)
-
-        elif transformation == "count":
+        if transformation == "count":
             self._encoders[column] = X[column].value_counts().to_dict()
 
-        elif transformation == "loo" and y is not None:
-            self._encoders[column] = LeaveOneOutEncoder(return_df=False).fit(
-                X[[column]], y
-            )
+        else:
+            encoder_class = encoder_classes.get(transformation)
 
-        elif transformation == "rankhot":
-            self._encoders[column] = RankHotEncoder().fit(X[[column]])
+            if callable(encoder_class):
+                self._encoders[column] = (
+                    encoder_class().fit(X[[column]], y)
+                    if y is not None
+                    else encoder_class().fit(X[[column]])
+                )
 
-    def _safe_transform(self, value, transformer, known_classes):
-        return transformer.transform([value])[0] if value in known_classes else -1
+            else:
+                self._encoders[column] = encoder_class().fit(X[[column]])
 
     def transform(self, X, y=None):
         X = X.copy()
-
         for column, transformation in self.encodings_options.items():
             X[column] = X[column].fillna("Unknown")
-
             if column in self._encoders:
-                transformer = self._encoders[column]
-                X = self._transform_column(X, column, transformation, transformer)
-
+                X = self._transform_column(
+                    X, column, transformation, self._encoders[column]
+                )
         return X
 
     def _transform_column(self, X, column, transformation, transformer):
-        if transformation == "label":
-            known_classes = set(transformer.classes_)
-            X[column] = X[column].apply(
-                lambda val: self._safe_transform(val, transformer, known_classes)
-            )
-
-        elif transformation in ["ordinal"]:
+        if transformation in ["label", "ordinal"]:
             X[column] = transformer.transform(X[[column]])
 
         elif transformation in [
-            "onehot",
-            "dummy",
+            "backward_diff",
+            "basen",
             "binary",
-            "target",
+            "catboost",
+            "dummy",
+            "glmm",
+            "gray",
+            "hashing",
+            "helmert",
+            "james_stein",
             "loo",
+            "m_estimate",
+            "onehot",
+            "polynomial",
+            "quantile",
             "rankhot",
+            "sum",
+            "target",
+            "woe",
         ]:
             encoded_array = transformer.transform(X[[column]])
             columns = transformer.get_feature_names_out([column])
-
-            if transformation in ["target", "loo"]:
-                columns = [f"{col}_{transformation}" for col in columns]
-
-            encoded_df = pd.DataFrame(
-                encoded_array,
-                columns=columns,
-                index=X.index,
-            )
+            encoded_df = pd.DataFrame(encoded_array, columns=columns, index=X.index)
             X = pd.concat([X.drop(columns=[column]), encoded_df], axis=1)
 
         elif transformation == "count":